backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
playwright==1.48.0
google-genai==0.3.0
google-cloud-storage==2.18.0
httpx==0.27.0
pydantic==2.9.0
python-multipart==0.0.12
